requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Redis
redis==5.0.1
hiredis==2.3.2

# NATS
nats-py==2.6.0

# Langfuse Observability
langfuse==2.20.0

# JSON Processing
orjson==3.9.12

# File Operations
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
validators==0.22.0

# Development
uvloop==0.19.0

